% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_adjust_dt_by_filetype.R
\name{adjust_dt_by_filetype}
\alias{adjust_dt_by_filetype}
\title{Process a LIFT CSV File and Correct Timestamps}
\usage{
adjust_dt_by_filetype(file_path, time_threshold = 1, time_offset = 0)
}
\arguments{
\item{file_path}{Character. Full path to the CSV file.}

\item{time_threshold}{Numeric. Maximum allowed difference in seconds before correction (default: 1).}

\item{time_offset}{difftime or numeric. Time shift to apply if correction needed (default: 0 seconds).}
}
\value{
NULL. File is overwritten if correction is applied.
}
\description{
Automatically detects the file type (_data, _fit, _spectr) and
applies the appropriate date/time correction.
}
\examples{
\dontrun{
adjust_dt_by_filetype("plot123~~2025_08_22__14_30_15_01_type_data.csv")
}
}
