% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_LIFT_settings.R
\name{lift_settings}
\alias{lift_settings}
\title{Create LIFT Experiment Settings}
\usage{
lift_settings(f = 427, ff = 840, sep = ",", liftmodel = "L008")
}
\arguments{
\item{f}{Integer. Number of flashlets (datapoints per transient). Default = 427.}

\item{ff}{Integer. Spectral range (number of points collected). Default = 840.}

\item{sep}{Character. Separator used in raw data file. Default = ",".}

\item{liftmodel}{Character. Identifier of excitation protocol. Default = "L008".}
}
\value{
A list containing the experimental parameters.
}
\description{
This function defines the basic parameters for analyzing
Light-Induced Fluorescence Transients (LIFT) data.
}
\examples{
settings <- lift_settings()
print(settings$f)
print(settings$liftmodel)

}
