% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_LIFT_redox_fit.R
\name{redox_fit}
\alias{redox_fit}
\title{Reduction and Reoxidation Fit Calculation}
\usage{
redox_fit(df, f, method = c("relative", "absolute"), regression = TRUE)
}
\arguments{
\item{df}{data.table or data.frame from lift_parse_metadata}

\item{f}{integer, number of excitation flashlets (e.g. 427)}

\item{method}{character vector, one or both of "relative"/"Keller" and "absolute"/"Zendonadi"}
}
\value{
data.table with added columns: Fo, Fm, Fq.Fm, Relative, Fr1, Fr2, Fr

Also can perform Log-Log Regression on LIFT Fluorescence Data
}
\description{
Calculates fluorescence reduction parameters (Fo, Fm, Fq/Fm, Relative)
and re-oxidation/relaxation parameters (Fr1, Fr2, Fr) from LIFT data.
}
