% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_LIFT_regression_fit.R, R/9_LIFT_redox_fit.R
\name{regression_fit}
\alias{regression_fit}
\title{Perform Log-Log Regression on LIFT Fluorescence Data}
\usage{
regression_fit(df, f, reorder_cols = TRUE)

regression_fit(df, f, reorder_cols = TRUE)
}
\arguments{
\item{df}{data.table or data.frame from lift_parse_metadata}

\item{f}{Integer, number of flashlets per transient (used for grouping).}

\item{reorder_cols}{Logical, whether to reorder columns (default: TRUE)}

\item{method}{character vector, one or both of "relative"/"Keller" and "absolute"/"Zendonadi"}

\item{Transient}{A data.table containing cleaned LIFT transient data.}
}
\value{
A data.table containing merged regression results for Group 1 and Group 2.

data.table with added columns: Fo, Fm, Fq.Fm, Relative, Fr1, Fr2, Fr

Also can perform Log-Log Regression on LIFT Fluorescence Data

The function regression_fit() calculates slopes, intercepts, and other regression parameters
for two groups of a LIFT transient dataset.

A data.table containing merged regression results for Group 1 and Group 2.
}
\description{
This function calculates slopes, intercepts, and other regression parameters
for two groups of a LIFT transient dataset.

Calculates fluorescence reduction parameters (Fo, Fm, Fq/Fm, Relative)
and re-oxidation/relaxation parameters (Fr1, Fr2, Fr) from LIFT data.
}
